commit of hotgym demo
